Create Pre Project Overview
Creating a prerequisite overview for MapADT: Hash Table project. It lists a summary of the project and topics I need to review before I can start writing code for the project. This helps me better understand what the project is asking me to do. It also helps me better understand the topics that are being taught in class.